chore: fix add group tab switching
